--- src/main/java/kitchenpos/application/TableService.java
@@ -1,58 +1,0 @@
-package kitchenpos.application;
-
-import kitchenpos.dao.OrderRepository;
-import kitchenpos.dao.OrderTableRepository;
-import kitchenpos.common.domain.OrderStatus;
-import kitchenpos.domain.OrderTable;
-import kitchenpos.dto.request.OrderTableRequest;
-import kitchenpos.dto.response.OrderTableResponse;
-import org.springframework.stereotype.Service;
-import org.springframework.transaction.annotation.Transactional;
-
-import java.util.Arrays;
-import java.util.List;
-import java.util.NoSuchElementException;
-
-@Transactional(readOnly = true)
-@Service
-public class TableService {
-    private final OrderRepository orderRepository;
-    private final OrderTableRepository orderTableRepository;
-
-    public TableService(final OrderRepository orderRepository, final OrderTableRepository orderTableRepository) {
-        this.orderRepository = orderRepository;
-        this.orderTableRepository = orderTableRepository;
-    }
-
-    @Transactional
-    public OrderTableResponse create(final OrderTableRequest request) {
-        return OrderTableResponse.of(orderTableRepository.save(request.toOrderTable()));
-    }
-
-    public List<OrderTableResponse> list() {
-        return OrderTableResponse.of(orderTableRepository.findAll());
-    }
-
-    @Transactional
-    public OrderTableResponse changeEmpty(Long orderTableId, Boolean empty) {
-        OrderTable orderTable = findById(orderTableId);
-        if (orderRepository.existsByOrderTableIdAndOrderStatusIn(
-                orderTableId, Arrays.asList(OrderStatus.COOKING, OrderStatus.MEAL))) {
-            throw new IllegalArgumentException("조리중이거나 식사중인 테이블이 있습니다.");
-        }
-        orderTable.changeEmpty(empty);
-        return OrderTableResponse.of(orderTable);
-    }
-
-    @Transactional
-    public OrderTableResponse changeNumberOfGuests(Long orderTableId, Integer numberOfGuests) {
-        OrderTable orderTable = findById(orderTableId);
-        orderTable.changeNumberOfGuests(numberOfGuests);
-        return OrderTableResponse.of(orderTable);
-    }
-
-    public OrderTable findById(Long id) {
-        return orderTableRepository.findById(id)
-                .orElseThrow(NoSuchElementException::new);
-    }
-}
