--- src/main/java/kitchenpos/dao/OrderLineItemRepository.java
@@ -1,12 +1,0 @@
-package kitchenpos.dao;
-
-import kitchenpos.domain.OrderLineItem;
-import org.springframework.data.jpa.repository.JpaRepository;
-
-import java.util.List;
-import java.util.Optional;
-
-public interface OrderLineItemRepository extends JpaRepository<OrderLineItem, Long> {
-
-    List<OrderLineItem> findAllByOrderId(Long orderId);
-}
