--- src/main/java/kitchenpos/dao/OrderTableRepository.java
@@ -1,15 +1,0 @@
-package kitchenpos.dao;
-
-import kitchenpos.domain.OrderTable;
-import org.springframework.data.jpa.repository.JpaRepository;
-
-import java.util.List;
-import java.util.Optional;
-
-public interface OrderTableRepository extends JpaRepository<OrderTable, Long> {
-
-    List<OrderTable> findAllByIdIn(List<Long> orderTableIds);
-
-    List<OrderTable> findAllByTableGroupId(Long tableGroupId);
-
-}
