--- src/main/java/kitchenpos/dto/request/OrderTableRequest.java
@@ -1,30 +1,0 @@
-package kitchenpos.dto.request;
-
-import kitchenpos.domain.OrderTable;
-
-public class OrderTableRequest {
-
-    private int numberOfGuests;
-    private boolean empty;
-
-    public OrderTableRequest(int numberOfGuests, boolean empty) {
-        this.numberOfGuests = numberOfGuests;
-        this.empty = empty;
-    }
-
-    public OrderTableRequest() {
-    }
-
-    public OrderTable toOrderTable() {
-        return new OrderTable(numberOfGuests, empty);
-    }
-
-    public int getNumberOfGuests() {
-        return numberOfGuests;
-    }
-
-    public boolean isEmpty() {
-        return empty;
-    }
-
-}
