--- src/main/java/kitchenpos/dto/request/TableGroupRequest.java
@@ -1,20 +1,0 @@
-package kitchenpos.dto.request;
-
-import java.util.List;
-
-public class TableGroupRequest {
-
-    private List<Long> orderTableIds;
-
-    public TableGroupRequest(List<Long> orderTableIds) {
-        this.orderTableIds = orderTableIds;
-    }
-
-    protected TableGroupRequest() {
-    }
-
-    public List<Long> getOrderTableIds() {
-        return orderTableIds;
-    }
-
-}
