+++ src/main/java/kitchenpos/order/domain/OrderLineItem.java
@@ -1,0 +1,54 @@
+package kitchenpos.order.domain;
+
+import javax.persistence.*;
+
+import static java.util.Objects.requireNonNull;
+
+@Entity
+public class OrderLineItem {
+
+    @Id
+    @GeneratedValue(strategy = GenerationType.IDENTITY)
+    private Long seq;
+    @ManyToOne
+    @JoinColumn(name = "order_id")
+    private Order order;
+    @Column(nullable = false)
+    private Long menuId;
+    private long quantity;
+
+    protected OrderLineItem() {
+    }
+
+    public OrderLineItem(Long seq, Order order, Long menuId, long quantity) {
+        this.seq = seq;
+        this.order = order;
+        this.menuId = requireNonNull(menuId, "menuId");
+        this.quantity = quantity;
+    }
+
+    public OrderLineItem(Long menuId, long quantity) {
+        this(null, null, menuId, quantity);
+    }
+
+    public void bindTo(Order order) {
+        this.order = order;
+    }
+
+    public Long getSeq() {
+        return seq;
+    }
+
+    public Order getOrder() {
+        return order;
+    }
+
+    public Long getMenuId() {
+        return menuId;
+    }
+
+    public long getQuantity() {
+        return quantity;
+    }
+}
+
