+++ src/main/java/kitchenpos/order/dto/OrderLineItemRequest.java
@@ -1,0 +1,30 @@
+package kitchenpos.order.dto;
+
+import kitchenpos.order.domain.OrderLineItem;
+
+public class OrderLineItemRequest {
+
+    private Long menuId;
+    private long quantity;
+
+    public OrderLineItemRequest(Long menuId, long quantity) {
+        this.menuId = menuId;
+        this.quantity = quantity;
+    }
+
+    protected OrderLineItemRequest() {
+    }
+
+    public OrderLineItem toOrderLineItem() {
+        return new OrderLineItem(menuId, quantity);
+    }
+
+    public Long getMenuId() {
+        return menuId;
+    }
+
+    public long getQuantity() {
+        return quantity;
+    }
+
+}
