+++ src/main/java/kitchenpos/order/dto/OrderResponse.java
@@ -1,0 +1,60 @@
+package kitchenpos.order.dto;
+
+import kitchenpos.order.domain.Order;
+
+import java.time.LocalDateTime;
+import java.util.List;
+import java.util.stream.Collectors;
+
+public class OrderResponse {
+
+    private Long id;
+    private Long orderTableId;
+    private String orderStatus;
+    private LocalDateTime orderedTime;
+    private List<OrderLineItemResponse> orderLineItems;
+
+    protected OrderResponse(Long id, Long orderTableId, String orderStatus, LocalDateTime orderedTime,
+                            List<OrderLineItemResponse> orderLineItems) {
+        this.id = id;
+        this.orderTableId = orderTableId;
+        this.orderStatus = orderStatus;
+        this.orderedTime = orderedTime;
+        this.orderLineItems = orderLineItems;
+    }
+
+    protected OrderResponse() {
+    }
+
+    public static OrderResponse of(Order order) {
+        return new OrderResponse(order.getId(), order.getOrderTableId(), order.getOrderStatus().name(), order.getOrderedTime(),
+                OrderLineItemResponse.of(order.getOrderLineItems()));
+    }
+
+    public static List<OrderResponse> of(List<Order> orders) {
+        return orders.stream()
+                .map(OrderResponse::of)
+                .collect(Collectors.toList());
+    }
+
+    public Long getId() {
+        return id;
+    }
+
+    public Long getOrderTableId() {
+        return orderTableId;
+    }
+
+    public String getOrderStatus() {
+        return orderStatus;
+    }
+
+    public LocalDateTime getOrderedTime() {
+        return orderedTime;
+    }
+
+    public List<OrderLineItemResponse> getOrderLineItems() {
+        return orderLineItems;
+    }
+
+}
