+++ src/main/java/kitchenpos/order/repository/OrderRepository.java
@@ -1,0 +1,14 @@
+package kitchenpos.order.repository;
+
+import kitchenpos.order.domain.Order;
+import kitchenpos.common.domain.OrderStatus;
+import org.springframework.data.jpa.repository.JpaRepository;
+
+import java.util.List;
+
+public interface OrderRepository extends JpaRepository<Order, Long> {
+
+    boolean existsByOrderTableIdAndOrderStatusIn(Long orderTableId, List<OrderStatus> orderStatuses);
+
+    boolean existsByOrderTableIdInAndOrderStatusIn(List<Long> orderTableIds, List<OrderStatus> orderStatuses);
+}
