+++ src/main/java/kitchenpos/order/ui/OrderRestController.java
@@ -1,0 +1,42 @@
+package kitchenpos.order.ui;
+
+import kitchenpos.order.application.OrderService;
+import kitchenpos.common.domain.OrderStatus;
+import kitchenpos.order.dto.OrderRequest;
+import kitchenpos.order.dto.OrderResponse;
+import org.springframework.http.ResponseEntity;
+import org.springframework.web.bind.annotation.*;
+
+import java.net.URI;
+import java.util.List;
+
+@RestController
+public class OrderRestController {
+    private final OrderService orderService;
+
+    public OrderRestController(final OrderService orderService) {
+        this.orderService = orderService;
+    }
+
+    @PostMapping("/api/orders")
+    public ResponseEntity<OrderResponse> create(@RequestBody OrderRequest request) {
+        final OrderResponse created = orderService.create(request);
+        final URI uri = URI.create("/api/orders/" + created.getId());
+        return ResponseEntity.created(uri)
+                .body(created);
+    }
+
+    @GetMapping("/api/orders")
+    public ResponseEntity<List<OrderResponse>> list() {
+        return ResponseEntity.ok()
+                .body(orderService.list());
+    }
+
+    @PutMapping("/api/orders/{orderId}/order-status")
+    public ResponseEntity<OrderResponse> changeOrderStatus(
+            @PathVariable Long orderId,
+            @RequestBody String orderStatus
+    ) {
+        return ResponseEntity.ok(orderService.changeOrderStatus(orderId, OrderStatus.valueOf(orderStatus)));
+    }
+}
