+++ src/main/java/kitchenpos/table/domain/OrderTable.java
@@ -1,0 +1,82 @@
+package kitchenpos.table.domain;
+
+import javax.persistence.*;
+import java.util.Objects;
+
+@Entity
+public class OrderTable {
+
+    @Id
+    @GeneratedValue(strategy = GenerationType.IDENTITY)
+    private Long id;
+    @ManyToOne
+    @JoinColumn(name = "table_group_id")
+    private TableGroup tableGroup;
+    private int numberOfGuests;
+    private boolean empty;
+
+    protected OrderTable() {
+    }
+
+    public OrderTable(TableGroup tableGroup, int numberOfGuests, boolean empty) {
+        this.tableGroup = tableGroup;
+        this.numberOfGuests = numberOfGuests;
+        this.empty = empty;
+    }
+
+    public OrderTable(Long id, TableGroup tableGroup, int numberOfGuests, boolean empty) {
+        this.id = id;
+        this.tableGroup = tableGroup;
+        this.numberOfGuests = numberOfGuests;
+        this.empty = empty;
+    }
+
+    public OrderTable(int numberOfGuests, boolean empty) {
+        this.numberOfGuests = numberOfGuests;
+        this.empty = empty;
+    }
+
+    public Long getId() {
+        return id;
+    }
+
+    public TableGroup getTableGroup() {
+        return tableGroup;
+    }
+
+    public int getNumberOfGuests() {
+        return numberOfGuests;
+    }
+
+    public boolean isEmpty() {
+        return empty;
+    }
+
+    public void changeEmpty(Boolean empty) {
+        if (Objects.nonNull(tableGroup)) {
+            throw new IllegalArgumentException("테이블 그룹이 존재합니다.");
+        }
+        this.empty = empty;
+    }
+
+    public void changeNumberOfGuests(Integer numberOfGuests) {
+        if (numberOfGuests < 0) {
+            throw new IllegalArgumentException("손님의 수는 음수가 될 수 없습니다.");
+        }
+
+        if (this.empty) {
+            throw new IllegalArgumentException("비어있는 테이블입니다.");
+        }
+
+        this.numberOfGuests = numberOfGuests;
+    }
+
+    public void bindTo(TableGroup tableGroup) {
+        this.tableGroup = tableGroup;
+        this.empty = false;
+    }
+
+    public void unBind() {
+        this.tableGroup = null;
+    }
+}
