+++ src/main/java/kitchenpos/table/dto/OrderTableRequest.java
@@ -1,0 +1,30 @@
+package kitchenpos.table.dto;
+
+import kitchenpos.table.domain.OrderTable;
+
+public class OrderTableRequest {
+
+    private int numberOfGuests;
+    private boolean empty;
+
+    public OrderTableRequest(int numberOfGuests, boolean empty) {
+        this.numberOfGuests = numberOfGuests;
+        this.empty = empty;
+    }
+
+    public OrderTableRequest() {
+    }
+
+    public OrderTable toOrderTable() {
+        return new OrderTable(numberOfGuests, empty);
+    }
+
+    public int getNumberOfGuests() {
+        return numberOfGuests;
+    }
+
+    public boolean isEmpty() {
+        return empty;
+    }
+
+}
