+++ src/main/java/kitchenpos/table/dto/TableGroupRequest.java
@@ -1,0 +1,20 @@
+package kitchenpos.table.dto;
+
+import java.util.List;
+
+public class TableGroupRequest {
+
+    private List<Long> orderTableIds;
+
+    public TableGroupRequest(List<Long> orderTableIds) {
+        this.orderTableIds = orderTableIds;
+    }
+
+    protected TableGroupRequest() {
+    }
+
+    public List<Long> getOrderTableIds() {
+        return orderTableIds;
+    }
+
+}
