+++ src/main/java/kitchenpos/table/ui/TableRestController.java
@@ -1,0 +1,51 @@
+package kitchenpos.table.ui;
+
+import kitchenpos.table.application.TableService;
+import kitchenpos.table.dto.OrderTableRequest;
+import kitchenpos.table.dto.OrderTableResponse;
+import org.springframework.http.ResponseEntity;
+import org.springframework.web.bind.annotation.*;
+
+import java.net.URI;
+import java.util.List;
+
+@RestController
+public class TableRestController {
+    private final TableService tableService;
+
+    public TableRestController(final TableService tableService) {
+        this.tableService = tableService;
+    }
+
+    @PostMapping("/api/tables")
+    public ResponseEntity<OrderTableResponse> create(@RequestBody OrderTableRequest request) {
+        final OrderTableResponse created = tableService.create(request);
+        final URI uri = URI.create("/api/tables/" + created.getId());
+        return ResponseEntity.created(uri)
+                .body(created);
+    }
+
+    @GetMapping("/api/tables")
+    public ResponseEntity<List<OrderTableResponse>> list() {
+        return ResponseEntity.ok()
+                .body(tableService.list());
+    }
+
+    @PutMapping("/api/tables/{orderTableId}/empty")
+    public ResponseEntity<OrderTableResponse> changeEmpty(
+            @PathVariable Long orderTableId,
+            @RequestBody Boolean empty
+    ) {
+        return ResponseEntity.ok()
+                .body(tableService.changeEmpty(orderTableId, empty));
+    }
+
+    @PutMapping("/api/tables/{orderTableId}/number-of-guests")
+    public ResponseEntity<OrderTableResponse> changeNumberOfGuests(
+            @PathVariable Long orderTableId,
+            @RequestBody Integer numberOfGuests
+    ) {
+        return ResponseEntity.ok()
+                .body(tableService.changeNumberOfGuests(orderTableId, numberOfGuests));
+    }
+}
